refactor shiftTemplates as timestamps
By refactoring shiftTemplates as timestamps, ISO 8601 strings (with no
timezone data), we don't require any of the helper functions for parsing
and manipulating the previous "hh:mm a" data structure.

Another major benefit is that it becomes clear now when a shift goes
over from one day into the next, because the date is included in the
timestamp.

The date will be set to Jan 1 1970 for all shiftTemplate timestamps.

The benefit in choosing not to include a timezone in the timestamp is
that the timestamp will always be interpreted as local time. This means
that we 5:00 pm will always be interpreted as 5:00 pm, whether or not it
is daylight savings time in the local area - this is the correct
behavior. We store the actual shift as UTC, with timezone data, because
the date and DST information is relevant here.