read day data in calendar using local time.
This fixes the error that would cause dragging a scheduled shift from
one cell to another to be scheduled a day later than intended. This was
occuring because we were setting the new date for the shift by reading
the destination date in UTC. If the UTC time had already crossed into
the next day, then we would be updating the date an additional day.

There is still a bug, where certain shifts that begin in one day and end
in the next are not being correctly updated.